requirements.txt
playwright
openai
pillow
python-dotenv
selenium 
webdriver-manager